pyproject.toml
[build-system]
requires = [
    "setuptools",
    "setuptools-scm",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "virtual-school-backend"
version = "0.1.0"
description = "Backend for virtual school on aiohttp"
readme = "README.md"
authors = [
    {name = "Andrey Shahov", email = "a.shahov.garlic@yandex.ru"},
]
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "aiohttp",
    "uvloop",
    "psycopg[binary, pool]",
    "python-dotenv",
    "pyjwt[crypto]",
]

[project.optional-dependencies]
dev = [
    "pip-tools",
]
test = [
    "pytest",
]
docs = [
]

[project.scripts]
virtual-school-backend = "virtual_school_backend.__main__:main"
init_db = "virtual_school_backend.scripts.init_db:main"

[tool.pip-tools]
generate-hashes = true
allow-unsafe = true
